Move HSA Population to new file, zero-pad HSA strings
